viam-module/jetson-yolo-detector/requirements.txt
viam-sdk>=0.73.1
typing-extensions
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=9.0.0
